+++ src/Checkout/Traits/EmulatesFillAttributes.php
@@ -35,7 +35,7 @@ trait EmulatesFillAttributes
         }
     }
 
-    private function fill($target, array $attributes)
+    protected function fill($target, array $attributes)
     {
         if (method_exists($target, 'fill')) {
             $target->fill($attributes);