+++ src/Foundation/Factories/OrderFactory.php
@@ -25,7 +25,8 @@ class OrderFactory extends BaseOrderFactory
     {
         $orderData = [
             'billpayer' => $checkout->getBillpayer()->toArray(),
-            'shippingAddress' => $checkout->getShippingAddress()->toArray()
+            'shippingAddress' => $checkout->getShippingAddress()->toArray(),
+            'shipping_method_id' => $checkout->getShippingMethodId(),
         ];
 
         $items = $this->convertCartItemsToDataArray($checkout->getCart());